Fix build in legacy golang version
