+++ infra/conf/shadowsocks_legacy.go
@@ -1,17 +1,13 @@
-//go:build go1.18
-
+//go:build !go1.18
 package conf
 
 import (
 	"strings"
 
 	"github.com/golang/protobuf/proto"
-	"github.com/sagernet/sing-shadowsocks/shadowaead_2022"
-	C "github.com/sagernet/sing/common"
 	"github.com/xtls/xray-core/common/protocol"
 	"github.com/xtls/xray-core/common/serial"
 	"github.com/xtls/xray-core/proxy/shadowsocks"
-	"github.com/xtls/xray-core/proxy/shadowsocks_2022"
 )
 
 func cipherFromString(c string) shadowsocks.CipherType {
@@ -49,40 +45,6 @@ type ShadowsocksServerConfig struct {
 }
 
 func (v *ShadowsocksServerConfig) Build() (proto.Message, error) {
-	if C.Contains(shadowaead_2022.List, v.Cipher) {
-		if len(v.Users) > 0 {
-			if v.Cipher == "" {
-				return nil, newError("shadowsocks 2022 (multi-user): missing server method")
-			}
-			if !strings.Contains(v.Cipher, "aes") {
-				return nil, newError("shadowsocks 2022 (multi-user): only blake3-aes-*-gcm methods are supported")
-			}
-
-			config := new(shadowsocks_2022.MultiUserServerConfig)
-			config.Method = v.Cipher
-
-			config.Key = v.Password
-			config.Network = v.NetworkList.Build()
-
-			for _, user := range v.Users {
-				if user.Cipher != "" {
-					return nil, newError("shadowsocks 2022 (multi-user): users must have empty method")
-				}
-				config.Users = append(config.Users, &shadowsocks_2022.User{
-					Key:   user.Password,
-					Email: user.Email,
-				})
-			}
-			return config, nil
-		}
-
-		config := new(shadowsocks_2022.ServerConfig)
-		config.Method = v.Cipher
-		config.Key = v.Password
-		config.Network = v.NetworkList.Build()
-		return config, nil
-	}
-
 	config := new(shadowsocks.ServerConfig)
 	config.Network = v.NetworkList.Build()
 
@@ -147,34 +109,9 @@ func (v *ShadowsocksClientConfig) Build() (proto.Message, error) {
 		return nil, newError("0 Shadowsocks server configured.")
 	}
 
-	if len(v.Servers) == 1 {
-		server := v.Servers[0]
-		if C.Contains(shadowaead_2022.List, server.Cipher) {
-			if server.Address == nil {
-				return nil, newError("Shadowsocks server address is not set.")
-			}
-			if server.Port == 0 {
-				return nil, newError("Invalid Shadowsocks port.")
-			}
-			if server.Password == "" {
-				return nil, newError("Shadowsocks password is not specified.")
-			}
-
-			config := new(shadowsocks_2022.ClientConfig)
-			config.Address = server.Address.Build()
-			config.Port = uint32(server.Port)
-			config.Method = server.Cipher
-			config.Key = server.Password
-			return config, nil
-		}
-	}
-
 	config := new(shadowsocks.ClientConfig)
 	serverSpecs := make([]*protocol.ServerEndpoint, len(v.Servers))
 	for idx, server := range v.Servers {
-		if C.Contains(shadowaead_2022.List, server.Cipher) {
-			return nil, newError("Shadowsocks 2022 accept no multi servers")
-		}
 		if server.Address == nil {
 			return nil, newError("Shadowsocks server address is not set.")
 		}
